hello_world_backend/requirements.txt
Django==3.1
djangorestframework==3.11.1
markdown==3.2.2
django-filter==2.3.0
django-cors-headers==3.4.0 
Pillow==7.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyHamcrest==2.0.2
freezegun==1.0.0